requirements.txt
pandas
selenium>=4.0.0
requests>=2.25.0
webdriver-manager>=3.8.0
lxml>=4.6.0
urllib3~=2.5.0
mysql-connector-python~=9.4.0
openpyxl~=3.1.5